+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [version 1.7.3] 2024-03-04
 
+- Better Sorting fallback
 - Set API Key directly from the Stock Chart
 
 ## [version 1.7.2] 2024-03-03